Fix test import error: move tests.py content into tests/__init__.py
Both esp/dbmail/tests.py (from PR #4383 on main) and esp/dbmail/tests/
(package directory from inbox feature) existed simultaneously, causing
Python to fail with "module incorrectly imported" since a file and
package can't share the same module name.

Moves the tests.py content (SendMailTest, ActionHandlerTest,
MessageRequestTest, CronmailImportTest) into tests/__init__.py and
removes the conflicting tests.py file.
